refactor(i18n): export i18n as a module with configurations and pre-set language
Move i18n to its own module with set configuration and locale instead of repeating the configuration
in every command file
