+++ commands/clips.js
@@ -1,19 +1,16 @@
-const { LOCALE } = require("../util/EvobotUtil");
-const i18n = require("i18n");
+const i18n = require("../util/i18n");
 const fs = require("fs");
-
-i18n.setLocale(LOCALE);
 
 module.exports = {
   name: "clips",
-  description: i18n.__('clips.description'),
+  description: i18n.__("clips.description"),
   execute(message) {
-    fs.readdir("./sounds", function(err, files) {
+    fs.readdir("./sounds", function (err, files) {
       if (err) return console.log("Unable to read directory: " + err);
 
       let clips = [];
 
-      files.forEach(function(file) {
+      files.forEach(function (file) {
         clips.push(file.substring(0, file.length - 4));
       });
