+++ commands/invite.js
@@ -1,11 +1,8 @@
-const { LOCALE } = require("../util/EvobotUtil");
-const i18n = require("i18n");
-
-i18n.setLocale(LOCALE);
+const i18n = require("../util/i18n");
 
 module.exports = {
   name: "invite",
-  description: i18n.__('invite.description'),
+  description: i18n.__("invite.description"),
   execute(message) {
     return message.member
       .send(
