+++ commands/nowplaying.js
@@ -1,10 +1,6 @@
 const { splitBar } = require("string-progressbar");
 const { MessageEmbed } = require("discord.js");
-
-const { LOCALE } = require("../util/EvobotUtil");
-const i18n = require("i18n");
-
-i18n.setLocale(LOCALE);
+const i18n = require("../util/i18n");
 
 module.exports = {
   name: "np",
