+++ commands/queue.js
@@ -1,8 +1,5 @@
 const { MessageEmbed } = require("discord.js");
-const { LOCALE } = require("../util/EvobotUtil");
-const i18n = require("i18n");
-
-i18n.setLocale(LOCALE);
+const i18n = require("../util/i18n");
 
 module.exports = {
   name: "queue",
