+++ commands/remove.js
@@ -1,6 +1,5 @@
-const { canModifyQueue, LOCALE } = require("../util/EvobotUtil");
-const i18n = require("i18n");
-i18n.setLocale(LOCALE);
+const { canModifyQueue } = require("../util/Util");
+const i18n = require("../util/i18n");
 
 const pattern = /^[0-9]{1,2}(\s*,\s*[0-9]{1,2})*$/;
