+++ commands/shuffle.js
@@ -1,11 +1,9 @@
-const { canModifyQueue, LOCALE } = require("../util/EvobotUtil");
-const i18n = require("i18n");
-
-i18n.setLocale(LOCALE);
+const { canModifyQueue } = require("../util/Util");
+const i18n = require("../util/i18n");
 
 module.exports = {
   name: "shuffle",
-  description: i18n.__('shuffle.description'),
+  description: i18n.__("shuffle.description"),
   execute(message) {
     const queue = message.client.queue.get(message.guild.id);
     if (!queue) return message.channel.send(i18n.__("shuffle.errorNotQueue")).catch(console.error);
@@ -18,6 +16,6 @@ module.exports = {
     }
     queue.songs = songs;
     message.client.queue.set(message.guild.id, queue);
-    queue.textChannel.send(i18n.__mf('shuffle.result', {author: message.author})).catch(console.error);
+    queue.textChannel.send(i18n.__mf("shuffle.result", { author: message.author })).catch(console.error);
   }
 };
