+++ commands/skip.js
@@ -1,7 +1,5 @@
-const { canModifyQueue, LOCALE } = require("../util/EvobotUtil");
-const i18n = require("i18n");
-
-i18n.setLocale(LOCALE);
+const { canModifyQueue } = require("../util/Util");
+const i18n = require("../util/i18n");
 
 module.exports = {
   name: "skip",
