+++ commands/stop.js
@@ -1,11 +1,9 @@
-const { canModifyQueue, LOCALE } = require("../util/EvobotUtil");
-const i18n = require("i18n");
-
-i18n.setLocale(LOCALE);
+const { canModifyQueue } = require("../util/Util");
+const i18n = require("../util/i18n");
 
 module.exports = {
   name: "stop",
-  description: i18n.__('stop.description'),
+  description: i18n.__("stop.description"),
   execute(message) {
     const queue = message.client.queue.get(message.guild.id);
